+++ src/commonMain/kotlin/uk/gov/hmrc/calculator/utils/taxcode/TaxCodeExtensions.kt
@@ -18,7 +18,9 @@ package uk.gov.hmrc.calculator.utils.taxcode
 import uk.gov.hmrc.calculator.exception.InvalidTaxCodeException
 import uk.gov.hmrc.calculator.model.Country
 import uk.gov.hmrc.calculator.model.taxcodes.NTCode
+import uk.gov.hmrc.calculator.model.taxcodes.ScottishTaxCode
 import uk.gov.hmrc.calculator.model.taxcodes.TaxCode
+import uk.gov.hmrc.calculator.utils.clarification.Clarification
 import uk.gov.hmrc.calculator.utils.toCountry
 import kotlin.jvm.JvmSynthetic
 
@@ -56,4 +58,17 @@ internal fun String.extractDoubleFromEmergencyTaxCode(): Double =
 internal fun TaxCode.getTrueTaxFreeAmount(): Double {
     val amount = this.taxFreeAmount
     return if (amount > 0) amount - 9 else 0.0
+}
+
+@Suppress("ComplexMethod")
+@JvmSynthetic
+internal fun TaxCode.getTaxCodeClarification(userPaysScottishTax: Boolean): Clarification? {
+    val clarification = when {
+        (this is ScottishTaxCode) && userPaysScottishTax -> Clarification.SCOTTISH_INCOME_APPLIED
+        (this is ScottishTaxCode) && !userPaysScottishTax -> Clarification.SCOTTISH_CODE_BUT_OTHER_RATE
+        (this !is ScottishTaxCode) && userPaysScottishTax -> Clarification.NON_SCOTTISH_CODE_BUT_SCOTTISH_RATE
+        else -> null
+    }
+
+    return clarification
 }