+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Allow Pension to go above `annualAllowance`.
 - Fix Student Loan internal calculation to use yearly wage instead of deducting pension.
 - Fix internal object so they're not accessible by jvm consumer.
+- Refactor Pension Contribution to take one amount parameter instead of two.
+- Added Clarification response list.
+- Added more validation to `PensionValidator`. 
+- Remove £9 from K and T tax code.
+- Refactor Pension Contribution/Student Loan constructor
+- Refactor Pension Contribution internal calculation to monthly instead of yearly.
 
 ## [2.10.2] - 2024-03-08Z